fix: adopt Editor component on main page
